Provide lang in title
Instead og setting the title of new REPL to win id, set it to language
name followed by id:<WINID>. This way it looks more like what would have
been set if opened normally in terminal, and we can still set REPL by
winID if we wanted. This was specifically done so the ctrl+enter
switching to editor will work.